+++ app/stores/WalletSettingsStore.js
@@ -7,8 +7,7 @@ import globalMessages from '../i18n/global-messages';
 import type { AssuranceMode } from '../types/transactionAssuranceTypes';
 import type {
   UpdateWalletResponse,
-  ChangeWalletPasswordResponse,
-  SetWalletPasswordResponse,
+  UpdateWalletPasswordResponse,
 } from '../api';
 
 export default class WalletSettingsStore extends Store {
@@ -19,9 +18,8 @@ export default class WalletSettingsStore extends Store {
   ];
 
   /* eslint-disable max-len */
-  @observable changeWalletPasswordRequest: Request<ChangeWalletPasswordResponse> = new Request(this.api.changeWalletPassword);
-  @observable setWalletPasswordRequest: Request<SetWalletPasswordResponse> = new Request(this.api.setWalletPassword);
   @observable updateWalletRequest: Request<UpdateWalletResponse> = new Request(this.api.updateWallet);
+  @observable updateWalletPasswordRequest: Request<UpdateWalletPasswordResponse> = new Request(this.api.updateWalletPassword);
   /* eslint-enable max-len */
 
   @observable walletFieldBeingEdited = null;
@@ -29,41 +27,32 @@ export default class WalletSettingsStore extends Store {
 
   setup() {
     const a = this.actions.walletSettings;
-    a.changeWalletPassword.listen(this._changeWalletPassword);
-    a.setWalletPassword.listen(this._setWalletPassword);
-    a.updateWalletAssuranceLevel.listen(this._updateWalletAssuranceLevel);
     a.startEditingWalletField.listen(this._startEditingWalletField);
     a.stopEditingWalletField.listen(this._stopEditingWalletField);
     a.cancelEditingWalletField.listen(this._cancelEditingWalletField);
     a.updateWalletField.listen(this._updateWalletField);
+    a.updateWalletPassword.listen(this._updateWalletPassword);
+    a.updateWalletAssuranceLevel.listen(this._updateWalletAssuranceLevel);
   }
 
   @action _updateWalletAssuranceLevel = async ({ assurance }: { assurance: AssuranceMode }) => {
     const activeWallet = this.stores.wallets.active;
     if (!activeWallet) return;
     const { id: walletId, name } = activeWallet;
-    await this.updateWalletRequest.execute({ walletId, name, assurance });
+    const wallet = await this.updateWalletRequest.execute({ walletId, name, assurance }).promise;
+    if (!wallet) return;
     await this.stores.wallets.walletsRequest.patch(result => {
-      const wallet = _.find(result, { id: walletId });
-      wallet.assurance = assurance;
+      const walletIndex = _.findIndex(result, { id: walletId });
+      result[walletIndex] = wallet;
     });
   };
 
-  @action _changeWalletPassword = async ({ walletId, oldPassword, newPassword }: {
-    walletId: string, oldPassword: string, newPassword: string,
+  @action _updateWalletPassword = async ({ walletId, oldPassword, newPassword }: {
+    walletId: string, oldPassword: ?string, newPassword: ?string,
   }) => {
-    await this.changeWalletPasswordRequest.execute({ walletId, oldPassword, newPassword });
+    await this.updateWalletPasswordRequest.execute({ walletId, oldPassword, newPassword });
     this.actions.dialogs.closeActiveDialog.trigger();
-    this.changeWalletPasswordRequest.reset();
-    this.stores.wallets.refreshWalletsData();
-  };
-
-  @action _setWalletPassword = async ({ walletId, password }: {
-    walletId: string, password: string,
-  }) => {
-    await this.setWalletPasswordRequest.execute({ walletId, password });
-    this.actions.dialogs.closeActiveDialog.trigger();
-    this.setWalletPasswordRequest.reset();
+    this.updateWalletPasswordRequest.reset();
     this.stores.wallets.refreshWalletsData();
   };
 
@@ -73,10 +62,11 @@ export default class WalletSettingsStore extends Store {
     const { id: walletId, name, assurance } = activeWallet;
     const walletData = { walletId, name, assurance };
     walletData[field] = value;
-    await this.updateWalletRequest.execute(walletData);
+    const wallet = await this.updateWalletRequest.execute(walletData).promise;
+    if (!wallet) return;
     await this.stores.wallets.walletsRequest.patch(result => {
-      const wallet = _.find(result, { id: walletId });
-      wallet[field] = value;
+      const walletIndex = _.findIndex(result, { id: walletId });
+      result[walletIndex] = wallet;
     });
   };