+++ features/step_definitions/wallets-steps.js
@@ -19,11 +19,12 @@ export default function () {
         daedalus.stores.wallets.refreshWalletsData().then(() => done(wallet));
       });
     }, defaultWalletKeyFilePath);
-
+    this.wallet = result.value;
+    // Add or set the wallets for this scenario
     if (this.wallets != null) {
-      this.wallets.push(result.value);
+      this.wallets.push(this.wallet);
     } else {
-      this.wallets = result.value;
+      this.wallets = [this.wallet];
     }
   });
 
@@ -34,11 +35,12 @@ export default function () {
         daedalus.stores.wallets.refreshWalletsData().then(() => done(wallet));
       });
     }, defaultWalletKeyFilePath);
-
+    this.wallet = result.value;
+    // Add or set the wallets for this scenario
     if (this.wallets != null) {
-      this.wallets.push(result.value);
+      this.wallets.push(this.wallet);
     } else {
-      this.wallets = result.value;
+      this.wallets = [this.wallet];
     }
   });
 
@@ -98,6 +100,18 @@ export default function () {
 
   this.When(/^I click on the create wallet button in add wallet dialog$/, function () {
     return this.waitAndClick('.WalletAddDialog .createWalletButton');
+  });
+
+  this.When(/^I have one wallet address$/, function () {
+    return this.client.waitForVisible('.generatedAddress-1');
+  });
+
+  this.When(/^I enter spending password "([^"]*)"$/, function (password) {
+    return this.client.setValue('.WalletReceive_spendingPassword input', password);
+  });
+
+  this.When(/^I click on the "Generate new address" button$/, function () {
+    return this.client.click('.generateAddressButton');
   });
 
   this.When(/^I click on the restore wallet button in add wallet dialog$/, function () {
@@ -292,6 +306,10 @@ export default function () {
     let transactionAmounts = await this.client.getText('.Transaction_amount');
     transactionAmounts = [].concat(transactionAmounts);
     expect(expectedData.amount).to.include(transactionAmounts[0]);
+  });
+
+  this.Then(/^I should see two wallet addresses$/, function () {
+    return this.client.waitForVisible('.generatedAddress-2');
   });
 
 };