+++ CHANGELOG.md
@@ -29,6 +29,7 @@ Changelog
 - Acceptance test for "Restore wallet with and without spending password" feature
 - Acceptance test for "Create wallet with spending password" feature
 - Acceptance test for "Send money from a wallet with spending password" feature
+- Acceptance test for "Generate wallet address" feature
 
 ### Fixes
 
@@ -44,6 +45,7 @@ Changelog
 - Added missing "Add wallet" button label translation key
 - Prevent "Delete wallet" dialog from closing until deletion is over
 - Improved error handling on "Set/Change wallet password" dialog
+- Improved API response errors handling
 
 ### Chores